+++ packages/@lwc/engine-core/src/framework/context.ts
@@ -21,3 +21,13 @@ export class ContextRequestEvent extends CustomEvent<{
         });
     }
 }
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export interface ContextRuntimeAdapter<T extends object> {
+    isServerSide: boolean;
+    component: object;
+    provideContext<T extends object>(contextVariety: T, providedContextSignal: Signal<unknown>): void;
+    consumeContext<T extends object>(
+      contextVariety: T,
+      contextProvidedCallback: ContextProvidedCallback,
+    ): void;
+}
