--- packages/@lwc/integration-karma/test/component/context/lifecycle-hooks/x/contextChild/contextChild.html
@@ -1,3 +1,0 @@
-<template>
-    <p>Test Child: {name}</p>
-</template>
