--- packages/@lwc/integration-karma/test/component/context/lifecycle-hooks/x/contextParent/contextParent.html
@@ -1,5 +1,0 @@
-<template>
-    <template lwc:if={showChild}>
-        <x-context-child></x-context-child>
-      </template>
-</template>
