--- packages/@lwc/integration-karma/test/component/context/lifecycle-hooks/x/contextParent/contextParent.js
@@ -1,12 +1,0 @@
-import { LightningElement, api } from 'lwc';
-import { createAtomicState } from 'x/stateManager';
-
-export default class ContextParent extends LightningElement {
-    @api state = createAtomicState();
-
-    @api hideChild = false;
-
-    get showChild() {
-        return !this.hideChild;
-    }
-}
