+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/child/child.html
@@ -1,0 +1,3 @@
+<template>
+    <x-grandchild></x-grandchild>
+</template>
