+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/child/child.js
@@ -1,0 +1,6 @@
+import { LightningElement, api } from 'lwc';
+import childContextFactory from 'x/childContext';
+
+export default class Child extends LightningElement {
+    @api context = childContextFactory(); 
+}
