+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/childContext/childContext.js
@@ -1,0 +1,4 @@
+import { defineContext } from 'test-state';
+import parentContextFactory from 'x/parentContext';
+
+export default defineContext(parentContextFactory);
