+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/grandchild/grandchild.html
@@ -1,0 +1,2 @@
+<template>
+</template>
