+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/grandchild/grandchild.js
@@ -1,0 +1,7 @@
+import { LightningElement, api } from 'lwc';
+import { defineContext } from 'test-state';
+import childContextFactory from 'x/childContext';
+
+export default class Grandchild extends LightningElement {
+    @api context = defineContext(childContextFactory)(); 
+}
