+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/parent/parent.html
@@ -1,0 +1,4 @@
+<template>
+  <x-child></x-child>
+  <x-child></x-child>
+</template>
