+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/parent/parent.js
@@ -1,0 +1,13 @@
+import { LightningElement, api } from 'lwc';
+import parentContextFactory from 'x/parentContext';
+
+export default class Parent extends LightningElement {
+    @api context = parentContextFactory('parent provided value');
+    @api disconnect;
+
+    disconnectedCallback() {
+        if (this.disconnect) {
+            this.disconnect(this);
+        }
+    }
+}
