+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/parentContext/parentContext.js
@@ -1,0 +1,4 @@
+import { defineContext } from 'test-state';
+
+export default defineContext();
+
