+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/root/root.html
@@ -1,0 +1,4 @@
+<template>
+    <x-parent></x-parent>
+    <x-grandchild></x-grandchild>
+</template>
