+++ packages/@lwc/integration-karma/test/component/context/lifecycle/x/root/root.js
@@ -1,0 +1,4 @@
+import { LightningElement } from 'lwc';
+
+export default class Root extends LightningElement {
+}
