--- packages/@lwc/integration-karma/test/component/context/updating/x/contextChild/contextChild.html
@@ -1,3 +1,0 @@
-<template>
-    <x-context-leaf></x-context-leaf>
-</template>
