--- packages/@lwc/integration-karma/test/component/context/updating/x/contextChild/contextChild.js
@@ -1,4 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export default class ContextLeaf extends LightningElement {
-}
