--- packages/@lwc/integration-karma/test/component/context/updating/x/contextLeaf/contextLeaf.html
@@ -1,3 +1,0 @@
-<template>
-    <p>Leaf context: {name}</p>
-</template>
