--- packages/@lwc/integration-karma/test/component/context/updating/x/contextLeaf/contextLeaf.js
@@ -1,10 +1,0 @@
-import { LightningElement, api } from 'lwc';
-import { consumeStateFactory } from 'x/stateManager';
-
-export default class ContextLeaf extends LightningElement {
-    @api state = consumeStateFactory();
-
-    get name() {
-        return this.state.value.context?.value?.name ?? 'not available';
-    }
-}
