--- packages/@lwc/integration-karma/test/component/context/updating/x/contextParent/contextParent.html
@@ -1,4 +1,0 @@
-<template>
-  <x-context-child></x-context-child>
-  <x-context-child></x-context-child>
-</template>
