--- packages/@lwc/integration-karma/test/component/context/updating/x/contextParent/contextParent.js
@@ -1,6 +1,0 @@
-import { LightningElement, api } from 'lwc';
-import { nameStateFactory } from 'x/stateManager';
-
-export default class ContextParent extends LightningElement {
-    @api state = nameStateFactory();
-}
